Even Liber Abbaci

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -1,18 +1,22 @@
 #include <stdio.h>
-
+/**
+ * main - Entry point
+ * Return: Return alwys 0
+ *
+ */
 int main(void)
 {
-    long long int a = 1, b = 2, c, sum = 0;
+	int a = 1, b = 2, c, sum = 0;
 
-    while (b <= 4000000)
-    {
-        if (b % 2 == 0)
-            sum += b;
-        c = a + b;
-        a = b;
-        b = c;
-    }
-    printf("%lld\n", sum);
-    return (0);
+	while (b <= 4000000)
+	{
+		if (b % 2 == 0)
+			sum += b;
+		c = a + b;
+		a = b;
+		b = c;
+	}
+	printf("%d\n", sum);
+	return (0);
 }
